# docs/swagger.yaml
basePath: /api/v1/cbesuperapp/cps_action
definitions:
...
  cbe-super-app-cps-action_internal_constants_dto_service_details.MinimumTransferUpdateRequest:
    properties:
      min_amount:
        type: integer
    type: object
...
  cbe-super-app-cps-action_internal_constants_dto_service_details.ServiceFeeDetailResponse:
    properties:
      aboveAmount:
        type: integer
      aboveServiceFee:
        type: integer
      cap:
        $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_types.Cap'
      cbeglentry:
        $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_types.GLEntry'
      cbeifbglentry:
        $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_types.GLEntry'
      cbeifbproductCodes:
        $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_types.ProductCodes'
      cbeproductCodes:
        $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_types.ProductCodes'
      createdAt:
        type: string
      deletedAt:
        type: string
      enabled:
        type: boolean
      id:
        type: string
      isDeleted:
        type: boolean
      key:
        type: string
      lastModifiedAt:
        type: string
      paymentType:
        type: string
      serviceCode:
        type: string
      serviceName:
        type: string
      serviceType:
        type: string
      tiers:
        items:
          $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_types.Tier'
        type: array
    type: object
...
  cbe-super-app-cps-action_internal_constants_dto_service_details.SingleMaxTransferRequest:
    properties:
      corporate_daily_cap:
        type: integer
      corporate_single_cap:
        type: integer
      individual_daily_cap:
        type: integer
      individual_single_cap:
        type: integer
    type: object
...
  cbe-super-app-cps-action_internal_constants_localization.StandardResponse:
    properties:
      data: {}
      message:
        type: string
      ok:
        type: boolean
      status:
        type: integer
      timestamp:
        type: string
    type: object
...
  internal_handlers_rest_http_service_details.paginatedMaximumTransferCapResponse:
    properties:
      docs:
        items:
          $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_dto_service_details.MaximumTransferCapResponse'
        type: array
      meta:
        $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_types.PaginationMeta'
    type: object
  internal_handlers_rest_http_service_details.paginatedMinimumTransferCapResponse:
    properties:
      docs:
        items:
          $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_dto_service_details.MinimumTransferCapResponse'
        type: array
      meta:
        $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_types.PaginationMeta'
    type: object
  internal_handlers_rest_http_service_details.paginatedServiceDetails:
    properties:
      docs:
        items:
          $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_model.ServiceDetails'
        type: array
      meta:
        $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_types.PaginationMeta'
    type: object
  internal_handlers_rest_http_service_details.paginatedServiceFeeResponse:
    properties:
      docs:
        items:
          $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_dto_service_details.ServiceFeeResponse'
        type: array
      meta:
        $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_types.PaginationMeta'
    type: object
host: localhost:8080
info:
  contact:
    email: contact@eaglelionsystems.com
    name: API Support
  description: API documentation for CPS Action service
  title: CPS Action API
  version: 1.0.0
paths:
...
  /maximum-transfer-caps:
    get:
      description: Retrieves all maximum transfer caps with pagination
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_handlers_rest_http_service_details.paginatedMaximumTransferCapResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
      summary: Get All Maximum Transfer Caps
      tags:
      - ServiceDetails
...
  /minimum-transfer-cap/{id}:
    put:
      consumes:
      - application/json
      description: Updates minimum transfer cap by service ID
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      - description: Minimum Transfer Update DTO
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_dto_service_details.MinimumTransferUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
      summary: Update Minimum Transfer Cap
      tags:
      - ServiceDetails
  /minimum-transfer-caps:
    get:
      description: Retrieves all minimum transfer caps with pagination
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_handlers_rest_http_service_details.paginatedMinimumTransferCapResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
      summary: Get All Minimum Transfer Caps
      tags:
      - ServiceDetails
...
  /service-fee-tire/{id}:
    delete:
      description: Deletes service fee tire by service ID
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
      summary: Delete Service Fee Tire
      tags:
      - ServiceDetails
  /service-fees:
    get:
      description: Retrieves all service fees with pagination
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_handlers_rest_http_service_details.paginatedServiceFeeResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
      summary: Get All Service Fees
      tags:
      - ServiceDetails
  /service-fees/{id}:
    get:
      description: Retrieves service fee detail by service ID
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_dto_service_details.ServiceFeeDetailResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
      summary: Get Service Fee Detail
      tags:
      - ServiceDetails
    put:
      consumes:
      - application/json
      description: Updates service fee detail by service ID
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      - description: Service Fee Detail DTO
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_dto_service_details.ServiceFeeDetailDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
      summary: Update Service Fee
      tags:
      - ServiceDetails
  /services:
    get:
      description: Retrieves all services with pagination
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_handlers_rest_http_service_details.paginatedServiceDetails'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
      summary: Get All Services
      tags:
      - ServiceDetails
  /single-max-transfer/{id}:
    put:
      consumes:
      - application/json
      description: Updates single max transfer by service ID
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      - description: Single Max Transfer DTO
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_dto_service_details.SingleMaxTransferRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
      summary: Update Single Max Transfer
      tags:
      - ServiceDetails
  /total-max-transfer-cap:
    put:
      consumes:
      - application/json
      description: Updates total max transfer cap
      parameters:
      - description: Total Max Transfer Update DTO
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_dto_service_details.TotalMaxTransferUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
      summary: Update Total Max Transfer Cap
      tags:
      - ServiceDetails
  /total-transfer-caps:
    get:
      description: Retrieves all total transfer caps
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_dto_service_details.TotalTransferCapResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
      summary: Get All Total Transfer Caps
      tags:
      - ServiceDetails
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"